Swiping behaviour on cards half implemented
Firebase crash analyzing and closing crashes

Instant data changes in all over the app. i.e addition deletion conversion of Contacts , Notes, Inquiries

chips selected color

followup text prediction

search bar refresh button and searching flow on back pressing.

+++ app/src/main/java/com/example/muzafarimran/lastingsales/sync/DataSenderAsync.java
@@ -17,12 +17,15 @@ import com.android.volley.toolbox.Volley;
 import com.example.muzafarimran.lastingsales.SessionManager;
 import com.example.muzafarimran.lastingsales.events.InquiryDeletedEventModel;
 import com.example.muzafarimran.lastingsales.events.ContactDeletedEventModel;
+import com.example.muzafarimran.lastingsales.events.LeadContactAddedEventModel;
+import com.example.muzafarimran.lastingsales.events.NoteAddedEventModel;
 import com.example.muzafarimran.lastingsales.listeners.PostExecuteListener;
 import com.example.muzafarimran.lastingsales.providers.models.LSCall;
 import com.example.muzafarimran.lastingsales.providers.models.LSContact;
 import com.example.muzafarimran.lastingsales.providers.models.LSInquiry;
 import com.example.muzafarimran.lastingsales.providers.models.LSNote;
 import com.example.muzafarimran.lastingsales.providers.models.TempFollowUp;
+import com.example.muzafarimran.lastingsales.utils.NetworkAccess;
 import com.example.muzafarimran.lastingsales.utils.PhoneNumberAndCallUtils;
 
 import org.json.JSONException;
@@ -99,27 +102,27 @@ public class DataSenderAsync {
                 @Override
                 protected Void doInBackground(Object... params) {
                     try {
-//                        if (NetworkAccess.isNetworkAvailable(mContext)) {
-                        if (sessionManager.isUserSignedIn()) {
-                            Log.d(TAG, "Syncing");
-                            Log.d(TAG, "Token : " + sessionManager.getLoginToken());
-                            Log.d(TAG, "user_id : " + sessionManager.getKeyLoginId());
-                            addContactsToServer();
-                            updateContactsToServer();
-                            addCallsToServer();
-                            addInquiriesToServer();
-                            updateInquiriesToServer();
-                            deleteInquiriesFromServer();
-                            addNotesToServer();
-                            updateNotesToServer();
-                            addFollowupsToServer();
-                            deleteContactsFromServer();
+                        if (NetworkAccess.isNetworkAvailable(mContext)) {
+                            if (sessionManager.isUserSignedIn()) {
+                                Log.d(TAG, "Syncing");
+                                Log.d(TAG, "Token : " + sessionManager.getLoginToken());
+                                Log.d(TAG, "user_id : " + sessionManager.getKeyLoginId());
+                                addContactsToServer();
+                                updateContactsToServer();
+                                addCallsToServer();
+                                addInquiriesToServer();
+                                updateInquiriesToServer();
+                                deleteInquiriesFromServer();
+                                addNotesToServer();
+                                updateNotesToServer();
+                                deleteNotesFromServer();
+                                addFollowupsToServer();
+                                deleteContactsFromServer();
 //                            if (NetworkAccess.isWifiConnected(mContext)) {
 //                                addRecordingToServer();
 //                            }
-                        }
-//                        }
-                        else {
+                            }
+                        } else {
                             Log.d(TAG, "SyncNoInternet");
                         }
                     } catch (Exception e) {
@@ -297,8 +300,10 @@ public class DataSenderAsync {
                     contact.setServerId(responseObject.getString("id"));
                     contact.setSyncStatus(SyncStatus.SYNC_STATUS_LEAD_UPDATE_SYNCED);
                     contact.save();
-                    Log.d(TAG, "onResponse : ServerID : " + responseObject.getString("id"));
-//                    }
+                    Log.d(TAG, "onResponse : ServerIDofNote : " + responseObject.getString("id"));
+                    LeadContactAddedEventModel mCallEvent = new LeadContactAddedEventModel();
+                    TinyBus bus = TinyBus.from(mContext);
+                    bus.post(mCallEvent);
                 } catch (JSONException e) {
                     e.printStackTrace();
                 }
@@ -607,11 +612,11 @@ public class DataSenderAsync {
                 try {
 //                    Log.d(TAG, "onErrorResponse: error.networkResponse.data: " + error.networkResponse.data);
                     Log.d(TAG, "onErrorResponse: error.networkResponse.statusCode: " + error.networkResponse.statusCode);
-                    if(error.networkResponse.statusCode == 412){
+                    if (error.networkResponse.statusCode == 412) {
                         JSONObject jObj = new JSONObject(new String(error.networkResponse.data));
                         int responseCode = jObj.getInt("responseCode");
                         String responseObject = jObj.getString("response");
-                        if(responseCode == 236){
+                        if (responseCode == 236) {
                             Log.d(TAG, "onErrorResponse: " + responseObject);
                             inquiry.delete();
                             InquiryDeletedEventModel mCallEvent = new InquiryDeletedEventModel();
@@ -675,20 +680,22 @@ public class DataSenderAsync {
             public void onErrorResponse(VolleyError error) {
                 error.printStackTrace();
                 Log.d(TAG, "onErrorResponse: CouldNotSyncAddNote");
-
                 try {
-                    JSONObject jObj = new JSONObject(new String(error.networkResponse.data));
-                    int responseCode = jObj.getInt("responseCode");
+                    if (error.networkResponse != null) {
+                        JSONObject jObj = new JSONObject(new String(error.networkResponse.data));
+                        int responseCode = jObj.getInt("responseCode");
 //                    Toast.makeText(getApplicationContext(), "response: "+response.toString(), Toast.LENGTH_LONG).show();
 //                    if (responseCode == 200) {
-                    JSONObject responseObject = jObj.getJSONObject("response");
-                    String description = responseObject.getString("description");
-                    String id = responseObject.getString("id");
-                    note.setServerId(id);
-                    note.setSyncStatus(SyncStatus.SYNC_STATUS_NOTE_ADDED_SYNCED);
-                    note.save();
-                    Log.d(TAG, "onResponse addNoteReSynced: " + description + " ServerNoteID : " + note.getServerId());
-
+                        JSONObject responseObject = jObj.getJSONObject("response");
+                        String description = responseObject.getString("description");
+                        String id = responseObject.getString("id");
+                        note.setServerId(id);
+                        note.setSyncStatus(SyncStatus.SYNC_STATUS_NOTE_ADDED_SYNCED);
+                        note.save();
+                        Log.d(TAG, "onResponse addNoteReSynced: " + description + " ServerNoteID : " + note.getServerId());
+                    } else {
+                        Log.d(TAG, "onErrorResponse: no response may be poor internet");
+                    }
                 } catch (Exception e) {
                     e.printStackTrace();
                 }
@@ -762,6 +769,98 @@ public class DataSenderAsync {
             public void onErrorResponse(VolleyError error) {
                 error.printStackTrace();
                 Log.d(TAG, "onErrorResponse: CouldNotSyncUpdateNote");
+            }
+        }) {
+        };
+        sr.setRetryPolicy(new DefaultRetryPolicy(
+                MY_TIMEOUT_MS,
+                MY_MAX_RETRIES,
+                DefaultRetryPolicy.DEFAULT_BACKOFF_MULT));
+        queue.add(sr);
+    }
+
+    private void deleteNotesFromServer() {
+
+        List<LSNote> notesList = null;
+        if (LSNote.count(LSNote.class) > 0) {
+            notesList = LSNote.find(LSNote.class, "sync_status = ?", SyncStatus.SYNC_STATUS_NOTE_DELETE_NOT_SYNCED);
+            Log.d(TAG, "deleteNotesFromServer: count : " + notesList.size());
+            for (LSNote oneNote : notesList) {
+                Log.d(TAG, "Found Notes");
+                deleteNoteFromServerSync(oneNote);
+            }
+        }
+    }
+
+
+    private void deleteNoteFromServerSync(final LSNote note) {
+        Log.d(TAG, "deleteNoteFromServerSync: ServerIdOfNote: " + note.getServerId());
+        currentState = PENDING;
+        final String BASE_URL = MyURLs.DELETE_NOTE;
+        Uri builtUri = Uri.parse(BASE_URL)
+                .buildUpon()
+                .appendPath("" + note.getContactOfNote().getServerId())
+                .appendPath("notes")
+                .appendPath("" + note.getServerId())
+                .appendQueryParameter("api_token", "" + sessionManager.getLoginToken())
+                .build();
+        final String myUrl = builtUri.toString();
+        StringRequest sr = new StringRequest(Request.Method.DELETE, myUrl, new Response.Listener<String>() {
+            @Override
+            public void onResponse(String response) {
+                Log.d(TAG, "onResponse() delete note called with: response = [" + response + "]");
+                try {
+                    JSONObject jObj = new JSONObject(response);
+                    // ignoring responseCode etc because server is sending wrong Json format
+//                    int responseCode = jObj.getInt("responseCode");
+//                    Log.d(TAG, "onResponse: DeleteInquiry: "+jObj.toString());
+//                    if (responseCode == 200) {
+//                        JSONObject responseObject = jObj.getJSONObject("response");
+                    note.delete();
+                    NoteAddedEventModel mNoteAdded = new NoteAddedEventModel();
+                    TinyBus bus = TinyBus.from(mContext);
+                    bus.post(mNoteAdded);
+//                    }
+                } catch (JSONException e) {
+                    e.printStackTrace();
+                }
+            }
+        }, new Response.ErrorListener() {
+            @Override
+            public void onErrorResponse(VolleyError error) {
+                error.printStackTrace();
+                Log.d(TAG, "onErrorResponse: CouldNotSyncDeleteNote");
+                try {
+                    if (error.networkResponse != null) {
+                        Log.d(TAG, "onErrorResponse: error.networkResponse.statusCode: " + error.networkResponse.statusCode);
+                        if (error.networkResponse.statusCode == 412) {
+                            JSONObject jObj = new JSONObject(new String(error.networkResponse.data));
+                            int responseCode = jObj.getInt("responseCode");
+                            String responseObject = jObj.getString("response");
+                            if (responseCode == 236) {
+                                Log.d(TAG, "onErrorResponse: " + responseObject);
+                                note.delete();
+                                NoteAddedEventModel mNoteAdded = new NoteAddedEventModel();
+                                TinyBus bus = TinyBus.from(mContext);
+                                bus.post(mNoteAdded);
+                                Log.d(TAG, "onErrorResponse: Note doesn't exist on server");
+                                Toast.makeText(mContext, "Note doesn't exist on server", Toast.LENGTH_SHORT).show(); //TODO delete before putting in production
+                            }
+                        } else if (error.networkResponse.statusCode == 404) {
+//                        note.delete();
+//                        Log.d(TAG, "onErrorResponse: Note doesn't exist on server deleted and from mobile.");
+//                        Toast.makeText(mContext, "Note doesn't exist on server deleted and from mobile.", Toast.LENGTH_SHORT).show(); //TODO delete before putting in production
+                        } else if (error.networkResponse.statusCode == 401) {
+                            // Doesn't Exist on server.
+                            note.delete();
+                        }
+                    } else {
+                        Log.d(TAG, "onErrorResponse: no response may be poor internet");
+                    }
+
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
             }
         }) {
         };
